requirements.txt
absl-py==0.2.2
astor==0.6.2
bleach==1.5.0
dill==0.2.7.1
edward==1.3.5
gast==0.2.0
grpcio==1.12.0
html5lib==0.9999999
Markdown==2.6.11
more-itertools==4.2.0
multiprocess==0.70.5
numpy==1.14.3
pandas==0.23.0
pathos==0.2.1
pox==0.2.3
ppft==1.6.4.7.1
protobuf==3.5.2.post1
python-dateutil==2.7.3
pytz==2018.4
six==1.11.0
tensorboard==1.6.0
tensorflow-gpu==1.6.0
termcolor==1.1.0
Werkzeug==0.14.1
gensim==3.4.0
